requirements.txt
Flask==2.3.2
PyMySQL==1.1.0
gunicorn==21.2.0
python-dotenv
